Consolidate schema definitions logic in the `_Definitions` class
Some common patterns are defined as methods to avoid duplication
(e.g. `create_reference_to_schema`, used to store a schema as a
definition and returns a `'definition-reference'` schema). A couple
existing methods on the `GenerateSchema` class are moved (and
renamed to better describe what they do) to the `_Definitions`
class.

Proper docstrings are added to attribute and methods.

Type hints are tweaked when necessary.